requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
plexapi
python-dateutil
tzdata
typing_extensions
httpx
packaging
websocket-client
python-multipart